requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
python-dotenv==0.20.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
fastapi==0.89.1
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.1
openai==0.26.4
packaging==23.0
pathy==0.10.1
preshed==3.0.8
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.5.0/pt_core_news_lg-3.5.0-py3-none-any.whl
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
starlette==0.22.0
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
wasabi==1.1.1
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
